分离配置文件，添加视频点击数功能，添加定时任务功能，修改Redis工具类

+++ src/main/java/net/xhalo/video/controllers/VideoController.java
@@ -71,7 +71,7 @@ public class VideoController {
     }
 
     @RequestMapping(value = "video-{videoId}.html")
-    public String getVideoById(@PathVariable(value = "videoId") Integer videoId, Model model) {
+    public String getVideoById(@PathVariable(value = "videoId") Long videoId, Model model) {
         Video result = null;
         result = videoService.getVideoById(videoId);
         if(result == null)
